backend-django/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
python-dotenv
django-cors-headers
djangorestframework-simplejwt
dj-database-url
Pillow
requests
